feat(vitals): full HOSxP Nurse Note port + combined Temp/Pulse chart
Previous VitalsTab used ipt_pregnancy_vital_sign which covered ~50% of
the HOSxP Nurse Note fields users expected. Switched the data source
to ipd_nurse_note (PK nurse_note_id) — the canonical form referenced
from HOSxPIPDPatientAdmitNurseNoteEntryFormUnit.pas.

Entry dialog now covers 13 sections / ~70 fields: datetime, core
vitals, biometric, physical exam, PV, IBP/CVP/ICP, EtCO2/scores/O2,
fluid in (14), fluid out (8), stool/urine, food+meds, notes.

Chart was rebuilt as SVG with pixel layout matching HOSxP: shared 4-row
day header (date / admit day / OP day / 4h slots) acts as the X-axis
for all panels below; Temp + Pulse share a single dual-Y panel with a
red 37C reference line; RR and BP panels align to the same day domain
so columns read vertically across all three strips.

Also: exclude present/ (tutorial deck, not deployable) from docker
build context and git; new cpd-persist-decimal-factors integration
test covering CPD score decimal persistence.

+++ src/db/tables/cpd-scores.ts
@@ -14,8 +14,13 @@ export const cpdScoresTable: TableDefinition = {
     { name: 'risk_level', type: 'string', maxLength: 10 },
     { name: 'recommendation', type: 'string', maxLength: 500, nullable: true },
     { name: 'factor_gravida', type: 'integer', nullable: true },
-    { name: 'factor_anc_count', type: 'integer', nullable: true },
-    { name: 'factor_ga_weeks', type: 'integer', nullable: true },
+    // factor_anc_count and factor_ga_weeks are DECIMAL because their
+    // evaluators in src/config/risk-levels.ts return 1.5 — Postgres rejects
+    // a numeric literal into INTEGER (SQLite is loose, so the bug only
+    // surfaced under pglite/postgres). See test
+    // tests/integration/cpd-persist-decimal-factors.test.ts.
+    { name: 'factor_anc_count', type: 'decimal', nullable: true },
+    { name: 'factor_ga_weeks', type: 'decimal', nullable: true },
     { name: 'factor_height_cm', type: 'decimal', nullable: true },
     { name: 'factor_weight_diff', type: 'decimal', nullable: true },
     { name: 'factor_fundal_ht', type: 'decimal', nullable: true },
